--- lib/src/utils/globals/mod.rs
@@ -1,12 +1,0 @@
-pub mod packages;
-pub use packages::PACKAGES;
-pub mod paths;
-pub use paths::PATHS;
-pub mod envs;
-pub use envs::ENVS;
-pub mod urls;
-pub use urls::URLS;
-pub mod spinners;
-pub use spinners::SPINNERS;
-pub mod versions;
-pub use versions::VERSIONS;
